Initial commit, make it compile for ESP32. Had to uncomment some code sections for now due to compiler bug.

+++ cmake/checks/atomic_check.cpp
@@ -2,7 +2,9 @@
 
 static int test()
 {
-    std::atomic<long long> x;
+    //std::atomic<long long> x;
+    // TODO: ESP32 does not have hardware support for 64bit atomics. Adding software support would be a cleaner solution.
+    std::atomic<long> x;
     return x;
 }
